D106-D108: 200+ project improvements — likely/unlikely, __printf, static const, dead code, fallthrough, narrowing
Batch A: Resource leak fix (exec.c elf_buf) + dead code removal (ssh_crypto.c 6 funcs, tpm_tis.c 2 duplicates, pfkey.c) + include guard fix (errno_ext.h)
Batch B: likely/unlikely annotations (65 allocation-failure branches across 9 files) + macro definition
Batch C: static const test tables (10 files) + fallthrough annotations (22 annotations across 9 files)
Batch E: Narrowing assignment fixes (16 changes in 10 files — swap.c, ext2.h, bufcache.c, ahci.c, page_cache.c, erofs.c, bpf_maps.c, fsck.c, syscall.c)
Batch F: __printf format annotations (30+ functions) + -Wformat=2 in Makefile
Batch D: Const correctness verified as already complete.

128 files changed, 439 insertions(+), 453 deletions(-)
Zero-warning build with -Wall -Wextra -Werror -Wformat=2

+++ src/fs/erofs.c
@@ -129,7 +129,7 @@ static int erofs_read_inode_raw(uint32_t nid, struct erofs_inode_extended *inode
 {
     if (!erofs_mounted) return -ENODEV;
     /* Inode table starts after superblock. Each inode is at a fixed offset. */
-    uint32_t inode_offset = sizeof(struct erofs_superblock) + (uint32_t)nid * sizeof(struct erofs_inode_extended);
+    size_t inode_offset = sizeof(struct erofs_superblock) + (size_t)nid * sizeof(struct erofs_inode_extended);
     if (inode_offset + sizeof(struct erofs_inode_extended) > erofs_data_size)
         return -EIO;
     memcpy(inode, erofs_data_base + inode_offset, sizeof(struct erofs_inode_extended));
